made separate folders for avatargroup
