+++ lib/components/Avatar/index.ts
@@ -1,0 +1,1 @@
+export * from './Avatar';
