+++ lib/components/AvatarGroup/index.ts
@@ -1,0 +1,1 @@
+export * from './AvatarGroup';
